+++ server/src/app.ts
@@ -202,12 +202,11 @@ function processTrackedDs(trackedDatasets) {
 	const nonblocking = trackedDatasets.filter(
 		ds => ds.init.status === 'nonblocking' || ds.init.status == 'recoverableError'
 	)
-	// if no dataset loaded successfully, assume that there may be something wrong
-	// with serverconfig and/or code, not with dataset js/ts files, and
-	// crash the server to trigger rollback
 	if (!done.length && !nonblocking.length) {
-		// this should trigger a rollback notification
-		throw `${serverconfig.URL}: there were no datasets that loaded successfully`
+		// if no dataset loaded successfully, and only if there are genome + dataset entries,
+		// then assume that there may be something wrong with serverconfig and/or code,
+		// not with dataset js/ts files, and crash the server to trigger rollback
+		if (trackedDatasets.length) throw `${serverconfig.URL}: there were no datasets that loaded successfully`
 	} else {
 		if (done.length) {
 			console.log(`\n--- these datasets finished loading ---`)