+++ server/src/termdb.scatter.js
@@ -305,14 +305,15 @@ async function colorAndShapeSamples(refSamples, cohortSamples, data, q) {
 			}
 		}
 		let i = 0
-		const shapes = order(result.shapeMap, q.shapeTW, data?.refs)
+
+		const shapes = Object.entries(result.shapeMap).sort((a, b) => a[0].localeCompare(b[0]))
+
 		for (const [category, value] of shapes) {
 			if ('shape' in value) continue
 			if (q.shapeTW.term.values?.[value.key]?.shape != undefined) value.shape = q.shapeTW.term.values?.[value.key].shape
 			else value.shape = i
 			i++
 		}
-
 		result.colorLegend = q.colorTW
 			? order(result.colorMap, q.colorTW, data.refs)
 			: [['Default', { sampleCount: cohortSamples.length, color: 'blue', key: 'Default' }]]
@@ -441,10 +442,7 @@ function order(map, tw, refs) {
 			return binA - binB
 		})
 	} else {
-		entries.sort((a, b) => {
-			if (a[1].sampleCount > b[1].sampleCount) return -1
-			else return 1
-		})
+		entries.sort((a, b) => a[0].localeCompare(b[0]))
 	}
 	return entries
 }