+++ shared/types/src/terms/numeric.ts
@@ -1,4 +1,4 @@
-import { MinBaseQ, BaseTW, TermValues, BaseTerm } from '../index.ts'
+import type { MinBaseQ, BaseTW, TermValues, BaseTerm } from '../index.ts'
 
 export type RawRegularBin = Partial<RegularNumericBinConfig> & { preferredBins?: string }
 
@@ -38,15 +38,13 @@ export type NumericTerm = BaseTerm & {
 	bins: PresetNumericBins
 	values?: TermValues
 	unit?: string
+	/** tailored color scale for this term, so that when the term is used for color gradient in scatter, this set of colors will be used by default */
+	continuousColorScale?: { minColor: string; maxColor: string }
 	/*densityNotAvailable?: boolean //Not used?
 	logScale?: string | number
 	max?: number
 	min?: number
-	name?: string
 	skip0forPercentile?: boolean
-	tvs?: Tvs
-	values?: TermValues
-	unit?: string
 	valueConversion?: ValueConversion*/
 }